+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 This is a list of manually mantained changes and updates for each version.
 
+## Version 1.3.4
+
+## Added
+
+#### Added deprecated definitions in package.json
+
+## Changed
+
+None
+
+## Deprecated
+
+None
+
+## Removed
+
+None
+
+***
+
 ## Version 1.3.3
 
 ## Added